Allow passing configuration into the main app
