+++ example/test/widget_test.dart
@@ -5,14 +5,13 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
+import 'package:amplitude_flutter_example/my_app.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:amplitude_flutter_example/main.dart';
-
 void main() {
   testWidgets('MyApp shows Send Event button', (WidgetTester tester) async {
-    await tester.pumpWidget(MyApp());
+    await tester.pumpWidget(const MyApp('API_KEY'));
 
     expect(find.byWidget(const Text('Send Event')), findsOneWidget);
   });
